+++ src/reportlab/__init__.py
@@ -1,9 +1,9 @@
 #Copyright ReportLab Europe Ltd. 2000-2023
 #see license.txt for license details
 __doc__="""The Reportlab PDF generation library."""
-Version = "4.0.8"
+Version = "4.0.9"
 __version__=Version
-__date__='20231211'
+__date__='20240108'
 
 import sys, os